Add ParserContext to ValueParser
A ParseContext object holds, the current location address, a symbol
table, and a delimiter character for expression parsing.  It removes
ValueParser::Locator callback from the ValueParser constructor.

A ParserContext is passed to a Functor object as an eval argument, to
add a hook while evaluating a function body to make a temporal
SymbolTable which has arguments binding.